Fix multiple indexes issue
