+++ src/metrics/metric_types.go
@@ -1,7 +1,10 @@
 package metrics
 
 import (
+	"fmt"
+	"github.com/newrelic/nri-postgresql/src/collection"
 	"reflect"
+	"strings"
 )
 
 // QueryDefinition holds the query and the unmarshall model
@@ -19,4 +22,62 @@ func (qd QueryDefinition) GetQuery() string {
 func (qd QueryDefinition) GetDataModels() interface{} {
 	ptr := reflect.New(reflect.ValueOf(qd.dataModels).Type())
 	return ptr.Interface()
+}
+
+func (qd *QueryDefinition) insertDatabaseNames(databases collection.DatabaseList) *QueryDefinition {
+	databaseList := ""
+	for database := range databases {
+		databaseList += `'` + database + `',`
+	}
+	databaseList = databaseList[0 : len(databaseList)-1]
+
+	qd.query = strings.Replace(qd.query, `%DATABASES%`, databaseList, 1)
+
+	return qd
+}
+
+func (qd *QueryDefinition) insertSchemaTables(schemaList collection.SchemaList) *QueryDefinition {
+	schemaTables := make([]string, 0)
+	for schema, tableList := range schemaList {
+		for table := range tableList {
+			schemaTables = append(schemaTables, fmt.Sprintf("'%s.%s'", schema, table))
+		}
+	}
+
+	if len(schemaTables) == 0 {
+		return nil
+	}
+
+	schemaTablesString := strings.Join(schemaTables, ",")
+
+	newTableDef := &QueryDefinition{
+		dataModels: qd.dataModels,
+		query:      strings.Replace(qd.query, `%SCHEMA_TABLES%`, schemaTablesString, 1),
+	}
+
+	return newTableDef
+}
+
+func (qd *QueryDefinition) insertSchemaTableIndexes(schemaList collection.SchemaList) *QueryDefinition {
+	schemaTableIndexes := make([]string, 0)
+	for schema, tableList := range schemaList {
+		for table, indexList := range tableList {
+			for _, index := range indexList {
+				schemaTableIndexes = append(schemaTableIndexes, fmt.Sprintf("'%s.%s.%s'", schema, table, index))
+			}
+		}
+	}
+
+	if len(schemaTableIndexes) == 0 {
+		return nil
+	}
+
+	schemaTableIndexString := strings.Join(schemaTableIndexes, ",")
+
+	newIndexDef := &QueryDefinition{
+		dataModels: qd.dataModels,
+		query:      strings.Replace(qd.query, `%SCHEMA_TABLE_INDEXES%`, schemaTableIndexString, 1),
+	}
+
+	return newIndexDef
 }
